requirements.txt
fastapi
mongoengine
pydantic
email-validator
httpx
python-jose[cryptography]
passlib
bcrypt
uvicorn
dotenv
requests
redis

pytest
mongomock
pytest-cov
fakeredis

black
isort
pylint
